test: extract opening suite, combine focus related tests
